+++ spine-cpp/include/spine/IkConstraint.h
@@ -74,13 +74,13 @@ namespace spine {
 
 		/// Adjusts the local rotation of the bone so the world position of the tip is as close to the target position as
 		/// possible. The target is specified in the world coordinate system.
-		static void apply(Skeleton &skeleton, BonePose &bone, float targetX, float targetY, bool compress, bool stretch, bool uniform, float mix);
+		static void apply(Skeleton &skeleton, BonePose &bone, float targetX, float targetY, bool compress, bool stretch, ScaleY scaleY, float mix);
 
 		/// Adjusts the parent and child bone rotations so the tip of the child is as close to the target position as
 		/// possible. The target is specified in the world coordinate system.
 		/// @param child A direct descendant of the parent bone.
 		static void apply(Skeleton &skeleton, BonePose &parent, BonePose &child, float targetX, float targetY, int bendDirection, bool stretch,
-						  bool uniform, float softness, float mix);
+						  ScaleY scaleY, float softness, float mix);
 
 	private:
 		Array<BonePose *> _bones;