+++ spine-cpp/include/spine/IkConstraintData.h
@@ -36,10 +36,37 @@
 #include <spine/ConstraintData.h>
 #include <spine/PosedData.h>
 #include <spine/IkConstraintPose.h>
+#include <string.h>
 
 namespace spine {
 	class BoneData;
 	class IkConstraint;
+
+	enum ScaleY {
+		ScaleY_None = 0,
+		ScaleY_Uniform,
+		ScaleY_Volume
+	};
+
+	inline ScaleY ScaleY_valueOf(const char *value) {
+		if (strcmp(value, "uniform") == 0)
+			return ScaleY_Uniform;
+		else if (strcmp(value, "volume") == 0)
+			return ScaleY_Volume;
+		else
+			return ScaleY_None;
+	}
+
+	inline const char *ScaleY_toString(ScaleY scaleY) {
+		switch (scaleY) {
+			case ScaleY_Uniform:
+				return "uniform";
+			case ScaleY_Volume:
+				return "volume";
+			default:
+				return "none";
+		}
+	}
 
 	class SP_API IkConstraintData : public ConstraintDataGeneric<IkConstraint, IkConstraintPose> {
 		friend class SkeletonBinary;
@@ -67,16 +94,16 @@ namespace spine {
 
 		void setTarget(BoneData &inValue);
 
-		/// When true and IkConstraintPose::getCompress() or IkConstraintPose::getStretch() is used, the bone is scaled on both the
-		/// X and Y axes.
-		bool getUniform();
+		/// Determines how BonePose::getScaleY() changes when IkConstraintPose::getCompress() or IkConstraintPose::getStretch()
+		/// sets BonePose::getScaleX().
+		ScaleY getScaleY();
 
-		void setUniform(bool uniform);
+		void setScaleY(ScaleY scaleY);
 
 	private:
 		Array<BoneData *> _bones;
 		BoneData *_target;
-		bool _uniform;
+		ScaleY _scaleY;
 	};
 }