+++ spine-cpp/src/spine/BoneData.cpp
@@ -35,7 +35,8 @@
 using namespace spine;
 
 BoneData::BoneData(int index, const String &name, BoneData *parent)
-	: PosedDataGeneric<BonePose>(name), _index(index), _parent(parent), _length(0), _color(0.61f, 0.61f, 0.61f, 1.0f), _icon(), _visible(true) {
+	: PosedDataGeneric<BonePose>(name), _index(index), _parent(parent), _length(0), _color(0.61f, 0.61f, 0.61f, 1.0f), _icon(), _iconSize(1),
+	  _iconRotation(0), _visible(true) {
 	assert(index >= 0);
 }
 
@@ -65,6 +66,22 @@ const String &BoneData::getIcon() {
 
 void BoneData::setIcon(const String &icon) {
 	this->_icon = icon;
+}
+
+float BoneData::getIconSize() {
+	return _iconSize;
+}
+
+void BoneData::setIconSize(float iconSize) {
+	_iconSize = iconSize;
+}
+
+float BoneData::getIconRotation() {
+	return _iconRotation;
+}
+
+void BoneData::setIconRotation(float iconRotation) {
+	_iconRotation = iconRotation;
 }
 
 bool BoneData::getVisible() {