+++ spine-godot/spine_godot/SpineConstant.h
@@ -93,6 +93,12 @@ public:
 		BlendMode_Screen
 	};
 
+	enum ScaleY {
+		ScaleY_None = 0,
+		ScaleY_Uniform,
+		ScaleY_Volume
+	};
+
 	enum UpdateMode {
 		UpdateMode_Process,
 		UpdateMode_Physics,
@@ -118,6 +124,7 @@ VARIANT_ENUM_CAST(SpineConstant::PositionMode)
 VARIANT_ENUM_CAST(SpineConstant::SpacingMode)
 VARIANT_ENUM_CAST(SpineConstant::RotateMode)
 VARIANT_ENUM_CAST(SpineConstant::BlendMode)
+VARIANT_ENUM_CAST(SpineConstant::ScaleY)
 VARIANT_ENUM_CAST(SpineConstant::UpdateMode)
 VARIANT_ENUM_CAST(SpineConstant::BoneMode)
 VARIANT_ENUM_CAST(SpineConstant::Physics)