+++ spine-godot/spine_godot/SpineIkConstraintData.cpp
@@ -35,8 +35,8 @@ void SpineIkConstraintData::_bind_methods() {
 	ClassDB::bind_method(D_METHOD("get_bones"), &SpineIkConstraintData::get_bones);
 	ClassDB::bind_method(D_METHOD("get_target"), &SpineIkConstraintData::get_target);
 	ClassDB::bind_method(D_METHOD("set_target", "v"), &SpineIkConstraintData::set_target);
-	ClassDB::bind_method(D_METHOD("get_uniform"), &SpineIkConstraintData::get_uniform);
-	ClassDB::bind_method(D_METHOD("set_uniform", "v"), &SpineIkConstraintData::set_uniform);
+	ClassDB::bind_method(D_METHOD("get_scale_y"), &SpineIkConstraintData::get_scale_y);
+	ClassDB::bind_method(D_METHOD("set_scale_y", "v"), &SpineIkConstraintData::set_scale_y);
 	ClassDB::bind_method(D_METHOD("get_setup_pose"), &SpineIkConstraintData::get_setup_pose);
 }
 
@@ -69,14 +69,14 @@ void SpineIkConstraintData::set_target(Ref<SpineBoneData> v) {
 	}
 }
 
-bool SpineIkConstraintData::get_uniform() {
-	SPINE_CHECK(get_spine_object(), false)
-	return get_spine_constraint_data()->getUniform();
+SpineConstant::ScaleY SpineIkConstraintData::get_scale_y() {
+	SPINE_CHECK(get_spine_object(), SpineConstant::ScaleY_None)
+	return (SpineConstant::ScaleY) get_spine_constraint_data()->getScaleY();
 }
 
-void SpineIkConstraintData::set_uniform(bool v) {
+void SpineIkConstraintData::set_scale_y(SpineConstant::ScaleY v) {
 	SPINE_CHECK(get_spine_object(), )
-	get_spine_constraint_data()->setUniform(v);
+	get_spine_constraint_data()->setScaleY((spine::ScaleY) v);
 }
 
 Ref<SpineIkConstraintPose> SpineIkConstraintData::get_setup_pose() {